feat: implement loading state for projects and refactor ProjectsSection to fetch projects with GitHub data
